CustomerAIV3
So...

My scene has the store design in it and the NavMesh has been set up, so that the customers can now walk around the store. The model characters have also been added in, this is so that I could work with an actual character, the customers will walk to the points via the CustomerAIV2 script.

I have started on the CustomerAIV3 script, this will hopefully allow for the customers to wonder around the store. This meaning, that they will go to random points that are close to them, rather than just going to points all over the map, but this would work as well.